Added an example use of Engine with a Component that causes a GameObject to randomly move through the world.

+++ engine/src/main/java/org/dragonskulle/components/IFixedUpdate.java
@@ -16,5 +16,5 @@ public interface IFixedUpdate {
      *
      * @param deltaTime Approximate time between calls to fixedUpdate
      */
-    void fixedUpdate(double deltaTime);
+    void fixedUpdate(float deltaTime);
 }
